Start reshuffling OME
