+++ src/eko/evolution_operator/grid.py
@@ -11,13 +11,12 @@ import numbers
 import numpy as np
 import numpy.typing as npt
 
-from ekore import matching_conditions
 from .. import member
 from .. import scale_variations as sv
 from ..io.runcards import Configs, Debug
-from ekore.matching_conditions.operator_matrix_element import OperatorMatrixElement
 from ..thresholds import flavor_shift, is_downward_path
-from . import Operator, flavors, physical
+from . import Operator, flavors, matching_condition, physical
+from .operator_matrix_element import OperatorMatrixElement
 
 logger = logging.getLogger(__name__)
 
@@ -245,7 +244,7 @@ class OperatorGrid(sv.ModeMixin):
 
             # join with the basis rotation, since matching requires c+ (or likewise)
             if is_downward:
-                matching = matching_conditions.MatchingCondition.split_ad_to_evol_map(
+                matching = matching_condition.MatchingCondition.split_ad_to_evol_map(
                     self._matching_operators[op.q2_to],
                     op.nf - 1,
                     op.q2_to,
@@ -256,7 +255,7 @@ class OperatorGrid(sv.ModeMixin):
                 )
                 final_op = final_op @ matching @ invrot @ phys_op
             else:
-                matching = matching_conditions.MatchingCondition.split_ad_to_evol_map(
+                matching = matching_condition.MatchingCondition.split_ad_to_evol_map(
                     self._matching_operators[op.q2_to],
                     op.nf,
                     op.q2_to,
