+++ src/ekore/operator_matrix_elements/as2.py
@@ -12,6 +12,7 @@ import numba as nb
 import numpy as np
 
 from eko import constants
+
 from ..harmonics.constants import zeta2, zeta3
 from .as1 import A_gg as A_gg_1
 from .as1 import A_hg as A_hg_1
