+++ src/ekore/operator_matrix_elements/as3/agg.py
