+++ src/ekore/operator_matrix_elements/as3/aggTF2.py
