+++ tests/eko/evolution_operator/test_physical.py
@@ -3,8 +3,8 @@ import pytest
 
 from eko import basis_rotation as br
 from eko import member
+from eko.evolution_operator.matching_condition import MatchingCondition
 from eko.evolution_operator.physical import PhysicalOperator
-from ekore.matching_conditions import MatchingCondition
 
 
 def mkOM(shape):
