+++ tests/ekore/operator_matrix_elements/test_nlo.py
@@ -2,7 +2,7 @@
 import numpy as np
 
 from ekore.harmonics import compute_cache
-from ekore.matching_conditions.as1 import A_ns, A_singlet
+from ekore.operator_matrix_elements.as1 import A_ns, A_singlet
 
 
 def test_A_1_intrinsic():
